feat(judicial-system): move "Enduropna mál" button to footer action
Moves the reopen case button from the top of the page (inside
CourtCaseInfo) to the footer action button in the Completed view.

When the case has been sent to the public prosecutor (canReopenCase),
the footer shows "Enduropna mál" (destructive). Otherwise it shows
"Leiðrétta mál" as before — the two are mutually exclusive.

Removes the button, modal, and all associated state from CourtCaseInfo
and relocates the full reopen-with-reason flow to Completed.tsx.
Tests moved from CaseInfo.spec.tsx to a new Completed.spec.tsx.
