+++ apps/judicial-system/web/src/components/CaseInfo/CaseInfo.spec.tsx
@@ -1,30 +1,11 @@
-import { render, screen } from '@testing-library/react'
-
 import {
-  AppealCaseState,
-  CaseIndictmentRulingDecision,
-  CaseState,
   CaseType,
   Defendant,
   Gender,
-  UserRole,
 } from '@island.is/judicial-system-web/src/graphql/schema'
 
-import { mockCase } from '../../utils/mocks'
-import {
-  IntlProviderWrapper,
-  UserContextWrapper,
-} from '../../utils/testHelpers'
 import { createFormatMessage } from '../../utils/testHelpers.logic'
-import { CourtCaseInfo, getDefendantLabel } from './CaseInfo'
-
-jest.mock('next/router', () => ({
-  useRouter: () => ({ push: jest.fn() }),
-}))
-
-jest.mock('@island.is/judicial-system-web/src/utils/hooks', () => ({
-  useCase: () => ({ updateCase: jest.fn() }),
-}))
+import { getDefendantLabel } from './CaseInfo'
 
 describe('getDefendantLabel - Indictment', () => {
   const formatMessage = createFormatMessage()
@@ -73,130 +54,3 @@ describe('getDefendantLabel - RestrictionCase/InvestigationCase', () => {
   })
 })
 
-describe('CourtCaseInfo - reopen button visibility', () => {
-  const completedIndictmentCase = {
-    ...mockCase(CaseType.INDICTMENT),
-    state: CaseState.COMPLETED,
-    indictmentCompletedDate: '2024-01-01',
-    indictmentSentToPublicProsecutorDate: '2024-01-02',
-  }
-
-  const renderComponent = (
-    userRole: UserRole,
-    workingCase = completedIndictmentCase,
-  ) =>
-    render(
-      <IntlProviderWrapper>
-        <UserContextWrapper userRole={userRole}>
-          <CourtCaseInfo workingCase={workingCase} />
-        </UserContextWrapper>
-      </IntlProviderWrapper>,
-    )
-
-  it('shows the reopen button to a district court user when there is no appeal', () => {
-    renderComponent(UserRole.DISTRICT_COURT_JUDGE)
-
-    expect(
-      screen.getByRole('button', { name: 'Enduropna mál' }),
-    ).toBeInTheDocument()
-  })
-
-  it('hides the reopen button from non-district-court users', () => {
-    renderComponent(UserRole.PROSECUTOR)
-
-    expect(
-      screen.queryByRole('button', { name: 'Enduropna mál' }),
-    ).not.toBeInTheDocument()
-  })
-
-  it('hides the reopen button when there is an active appeal', () => {
-    const caseWithActiveAppeal = {
-      ...completedIndictmentCase,
-      appealCase: { id: 'appeal_id', appealState: AppealCaseState.RECEIVED },
-    }
-
-    renderComponent(UserRole.DISTRICT_COURT_JUDGE, caseWithActiveAppeal)
-
-    expect(
-      screen.queryByRole('button', { name: 'Enduropna mál' }),
-    ).not.toBeInTheDocument()
-  })
-
-  it('hides the reopen button when the appeal is in APPEALED state', () => {
-    const caseWithAppealedAppeal = {
-      ...completedIndictmentCase,
-      appealCase: { id: 'appeal_id', appealState: AppealCaseState.APPEALED },
-    }
-
-    renderComponent(UserRole.DISTRICT_COURT_JUDGE, caseWithAppealedAppeal)
-
-    expect(
-      screen.queryByRole('button', { name: 'Enduropna mál' }),
-    ).not.toBeInTheDocument()
-  })
-
-  it('shows the reopen button when the appeal is completed', () => {
-    const caseWithCompletedAppeal = {
-      ...completedIndictmentCase,
-      appealCase: { id: 'appeal_id', appealState: AppealCaseState.COMPLETED },
-    }
-
-    renderComponent(UserRole.DISTRICT_COURT_JUDGE, caseWithCompletedAppeal)
-
-    expect(
-      screen.getByRole('button', { name: 'Enduropna mál' }),
-    ).toBeInTheDocument()
-  })
-
-  it('shows the reopen button when the appeal is withdrawn', () => {
-    const caseWithWithdrawnAppeal = {
-      ...completedIndictmentCase,
-      appealCase: { id: 'appeal_id', appealState: AppealCaseState.WITHDRAWN },
-    }
-
-    renderComponent(UserRole.DISTRICT_COURT_JUDGE, caseWithWithdrawnAppeal)
-
-    expect(
-      screen.getByRole('button', { name: 'Enduropna mál' }),
-    ).toBeInTheDocument()
-  })
-
-  it('hides the reopen button when the indictment ruling decision is withdrawal', () => {
-    const withdrawnCase = {
-      ...completedIndictmentCase,
-      indictmentRulingDecision: CaseIndictmentRulingDecision.WITHDRAWAL,
-    }
-
-    renderComponent(UserRole.DISTRICT_COURT_JUDGE, withdrawnCase)
-
-    expect(
-      screen.queryByRole('button', { name: 'Enduropna mál' }),
-    ).not.toBeInTheDocument()
-  })
-
-  it('hides the reopen button when the case has been merged', () => {
-    const mergedCase = {
-      ...completedIndictmentCase,
-      mergeCase: { id: 'merged_case_id' },
-    }
-
-    renderComponent(UserRole.DISTRICT_COURT_JUDGE, mergedCase)
-
-    expect(
-      screen.queryByRole('button', { name: 'Enduropna mál' }),
-    ).not.toBeInTheDocument()
-  })
-
-  it('hides the reopen button when the case has not been sent to public prosecutor', () => {
-    const notSentCase = {
-      ...completedIndictmentCase,
-      indictmentSentToPublicProsecutorDate: undefined,
-    }
-
-    renderComponent(UserRole.DISTRICT_COURT_JUDGE, notSentCase)
-
-    expect(
-      screen.queryByRole('button', { name: 'Enduropna mál' }),
-    ).not.toBeInTheDocument()
-  })
-})
